bfcp: fix win32 warning

+++ src/bfcp/attr.c
@@ -128,7 +128,7 @@ static int attr_encode(struct mbuf *mb, bool mand, enum bfcp_attrib type,
 
 	mb->pos = start;
 	err |= mbuf_write_u8(mb, (type<<1) | (mand ? 1 : 0));
-	err |= mbuf_write_u8(mb, len);
+	err |= mbuf_write_u8(mb, (uint8_t)len);
 	mb->pos += (len - BFCP_ATTR_HDR_SIZE);
 
 	/* padding */
